Remove callbacks module
Due to an issue related to using a callbacks class with after commit.
