+++ spec/dummy/app/models/balance.rb
@@ -3,5 +3,5 @@
 class Balance < ActiveRecord::Base
   belongs_to :user
 
-  around_create ::Eventish::Callback
+  around_create -> { ::Eventish.publish('balance_around_create', self) }
 end
