+++ spec/dummy/config/initializers/eventish.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'eventish/active_job_event'
-require 'eventish/callback'
 require 'eventish/adapters/active_support'
 
 Eventish.setup do |config|
