+++ spec/dummy/config/initializers/test_event.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 Rails.configuration.after_initialize do
-  Eventish.adapter.publish('app_loaded')
+  Eventish.publish('app_loaded')
 end
